log: write to stderr
Write errors to standard error instead of syslog so the supervising
process can capture the output and route accordingly.

Convert the logging to use a variadic function to avoid formatting the
messages to a temporary buffer.

TODO: LogMsg() is called from the signal handler and is not async signal
safe.